Mandatory messages + spaces.
LoginScreen:
In line 99, a message was added to let the user know that feilds are mandatory.
And a mandatoryTextStyle CSS class was added.

SignUpScreen:
In line 29, trimmed spaces from displayname and passwords.
Line 52, set displayname without spcaes.
Line 123a message was added to let the user know that feilds are mandatory.
And a mandatoryTextStyle CSS class was added.